[hotfix][runtime] It was possible to traverse the directory of the host through /jobmanager/logs/<path-to-file>.
The <path-to-file> had to be modified in a way that '../' referring to the parent folder needed to be replaced by '..%252f'. This enabled traversing the directory structure relative to the ./logs folder, e.g. /jobmanager/logs/..%252f/README.txt would return the content of the README.txt located in the FLINK_HOME folder (assuming Flink's default folder structure).

This is fixed now. The passed path is ignored in the same way as it's already done for the TaskManagerCustomLogHandler.

(cherry picked from commit b561010b0ee741543c3953306037f00d7a9f0801)